# Small speed-up in BetweenBB