This is a non-functional speed-up.  I did not think it needs testing on the framework. 
100 runs of bench on my machine yields:
master
10,791,432 cycles
15,871,536,830 instructions
3.074 seconds

patch
10,736,677,936 cycles
15,834,958,543 instructions
3.055 seconds.