@vondele 2020-03-05
for reference : https://godbolt.org/z/zwuTMM

so, yes, nicer code being generated.